requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
django-su==1.0.0
django-taggit==3.1.0
djangorestframework==3.14.0
Pillow==9.3.0
djoser @ git+https://github.com/sunscrapers/djoser.git@abdf622f95dfa2c6278c4bd6d50dfe69559d90c0
